render-requirements.txt
gunicorn==20.1.*
flask==2.3.*
flask-sqlalchemy==3.0.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
pandas==1.5.*
flask-restful==0.3.*
flask-cors==3.0.*
flask-limiter==2.8.*
flask-wtf==1.0.*
